// src/api/openapi/auto-generated/models/GroupStrategySettings.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GroupStrategySettings
 */
export interface GroupStrategySettings {
    /**
     * Strategies and their values for the group.
     * @type {{ [key: string]: string; }}
     * @memberof GroupStrategySettings
     */
    strategies: { [key: string]: string |undefined; };
}

/**
 * Check if a given object implements the GroupStrategySettings interface.
 */
export function instanceOfGroupStrategySettings(value: object): value is GroupStrategySettings {
    if (!('implementationYear' in value) || value['implementationYear'] === undefined) return false;
    if (!('strategies' in value) || value['strategies'] === undefined) return false;
    return true;
}

export function GroupStrategySettingsFromJSON(json: any): GroupStrategySettings {
    return GroupStrategySettingsFromJSONTyped(json, false);
}

export function GroupStrategySettingsFromJSONTyped(json: any, ignoreDiscriminator: boolean): GroupStrategySettings {
    if (json == null) {
        return json;
    }
    return {
        
        'implementationYear': json['implementation_year'],
        'strategies': json['strategies'],
    };
}

export function GroupStrategySettingsToJSON(value?: GroupStrategySettings | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'implementation_year': value['implementationYear'],
        'strategies': value['strategies'],
    };
}